+++ src/controllers/person.ctrl.js
@@ -2,7 +2,7 @@ const driver = require("../config/db-driver");
 const helpers = require("../helpers");
 
 class Person {
-  constructor({_id=null,label=null,honorificPrefix=[],firstName=null,middleName=null,lastName=null,fnameSoundex=null,lnameSoundex=null,alternateAppelations=[],description=null,status=false,createdBy=null,createdAt=null,updatedBy=null,updatedAt=null}) {
+  constructor({_id=null,label=null,honorificPrefix=[],firstName=null,middleName=null,lastName=null,fnameSoundex=null,lnameSoundex=null,alternateAppelations=[],description=null,status='private',createdBy=null,createdAt=null,updatedBy=null,updatedAt=null}) {
     if (typeof _id!=="undefined" && _id!==null) {
       this._id = _id;
     }
@@ -58,19 +58,19 @@ class Person {
   validate() {
     let status = true;
     let errors = [];
-    if (this.firstName.length<2) {
+    if (this.firstName!==null && this.firstName.length<2) {
       status = false;
       errors.push({field: "firstName", msg: "First name must contain at least 2 characters"});
     }
-    if (this.middleName!=="" && this.firstName.length<2) {
+    if (this.firstName!==null && this.middleName!=="" && this.firstName.length<2) {
       status = false;
       errors.push({field: "middleName", msg: "If middle name is entered it must contain at least 2 characters"});
     }
-    if (this.lastName.length<2) {
+    if (this.firstName!==null && this.lastName.length<2) {
       status = false;
       errors.push({field: "lastName", msg: "Last name must contain at least 2 characters"});
     }
-    if (this.alternateAppelations.length>0) {
+    if (this.firstName!==null && this.alternateAppelations.length>0) {
       for (let key in this.alternateAppelations) {
         let alternateAppelation = this.alternateAppelations[key];
         let label = "";
@@ -227,6 +227,34 @@ class Person {
   }
 };
 
+/**
+* @api {get} /people Get people
+* @apiName get people
+* @apiGroup People
+*
+* @apiParam {string} [label] A string to match against the peoples' labels.
+* @apiParam {string} [firstName] A string to match against the peoples' first names.
+* @apiParam {string} [lastName] A string to match against the peoples' last names.
+* @apiParam {string} [fnameSoundex] A string to match against the peoples' first name soundex.
+* @apiParam {string} [lnameSoundex] A string to match against the peoples' last name soundex.
+* @apiParam {string} [description] A string to match against the peoples' description.
+* @apiParam {number} [page=1] The current page of results
+* @apiParam {number} [limit=25] The number of results per page
+* @apiSuccessExample {json} Success-Response:
+{
+  "status": true,
+  "data": {
+    "currentPage": 1,
+    "data": [
+      {"lastName": "Fox", "firstName": "Aidan", "honorificPrefix": [""], "middleName": "", "label": "Aidan Fox",…},
+    …],
+    "totalItems": "221",
+    "totalPages": 9
+  },
+  "error": [],
+  "msg": "Query results"
+}
+*/
 const getPeople = async (req, resp) => {
   let parameters = req.query;
   let label = "";
@@ -234,7 +262,6 @@ const getPeople = async (req, resp) => {
   let lastName = "";
   let fnameSoundex = "";
   let lnameSoundex = "";
-  let _id = "";
   let description = "";
   let page = 0;
   let queryPage = 0;
@@ -410,15 +437,25 @@ const getPeopleQuery = async (query, queryParams, limit) => {
   return result;
 }
 
+/**
+* @api {get} /person Get person
+* @apiName get person
+* @apiGroup People
+*
+* @apiParam {string} _id The _id of the requested person.
+* @apiSuccessExample {json} Success-Response:
+{"status":true,"data":{"_id":"2069","honorificPrefix":["My"],"firstName":"fname","middleName":"mname","lastName":"lname","label":"fname mname lname","fnameSoundex":"F550","lnameSoundex":"L550","description":"description","status":"private","alternateAppelations":[{"appelation":"","firstName":"altfname","middleName":"altmname","lastName":"altlname","note":"note","language":{"value":"en","label":"English"}}],"createdBy":"260","createdAt":"2020-01-14T15:39:10.638Z","updatedBy":"260","updatedAt":"2020-01-14T15:42:42.939Z","events":[],"organisations":[],"people":[],"resources":[]},"error":[],"msg":"Query results"}
+*/
 const getPerson = async(req, resp) => {
   let parameters = req.query;
-  if (typeof parameters._id==="undefined" && parameters._id==="") {
+  if (typeof parameters._id==="undefined" || parameters._id==="") {
     resp.json({
       status: false,
       data: [],
       error: true,
       msg: "Please select a valid id to continue.",
     });
+    return false;
   }
   let _id = parameters._id;
   let person = new Person({_id: _id});
@@ -431,8 +468,40 @@ const getPerson = async(req, resp) => {
   });
 }
 
+/**
+* @api {put} /person Put person
+* @apiName put person
+* @apiGroup People
+* @apiPermission admin
+*
+* @apiParam {string} [_id] The _id of the person. This should be undefined|null|blank in the creation of a new person.
+* @apiParam {array} [honorificPrefix] The various honorific prefixes a person has.
+* @apiParam {string} firstName The person's first name.
+* @apiParam {string} [middleName] The person's middle name.
+* @apiParam {string} lastName The person's lastName name.
+* @apiParam {string} [description] A description about the person.
+* @apiParam {string} [status=private] The status of the person.
+* @apiParam {array} [alternateAppelations] The person's alternate appelations.
+* @apiParam {string}  [alternateAppelation[appelation]] The person's alternate appelation label.
+* @apiParam {string}  alternateAppelation[firstName] The person's alternate appelation first name.
+* @apiParam {string}  [alternateAppelation[middleName]] The person's alternate appelation middle name.
+* @apiParam {string}  alternateAppelation[lastName] The person's alternate appelation lastName name.
+* @apiParam {string}  [alternateAppelation[note]] The person's alternate appelation note.
+* @apiParam {object}  [alternateAppelation[language]] The person's alternate appelation language.
+* @apiSuccessExample {json} Success-Response:
+{"error":[],"status":true,"data":{"lastName":"lname","updatedBy":"260","description":"description","honorificPrefix":["Mr"],"label":"fname mname lname","alternateAppelations":[],"firstName":"fname","createdAt":"2020-01-14T15:39:10.638Z","createdBy":"260","middleName":"mname","lnameSoundex":"L550","fnameSoundex":"F550","status":"private","updatedAt":"2020-01-14T15:39:10.638Z","_id":"2069"}}
+*/
 const putPerson = async(req, resp) => {
   let postData = req.body;
+  if (Object.keys(postData).length===0) {
+    resp.json({
+      status: false,
+      data: [],
+      error: true,
+      msg: "The person must not be empty",
+    });
+    return false;
+  }
   let personData = {};
   for (let key in postData) {
     if (postData[key]!==null) {
@@ -459,13 +528,67 @@ const putPerson = async(req, resp) => {
   resp.json(output);
 }
 
+/**
+* @api {delete} /person Delete person
+* @apiName delete person
+* @apiGroup People
+* @apiPermission admin
+*
+* @apiParam {string} _id The id of the person for deletion.
+*
+* @apiSuccessExample {json} Success-Response:
+{"status":true,"data":{"records":[],"summary":{"statement":{"text":"MATCH (n:Person) WHERE id(n)=2069 DELETE n","parameters":{}},"statementType":"w","counters":{"_stats":{"nodesCreated":0,"nodesDeleted":1,"relationshipsCreated":0,"relationshipsDeleted":0,"propertiesSet":0,"labelsAdded":0,"labelsRemoved":0,"indexesAdded":0,"indexesRemoved":0,"constraintsAdded":0,"constraintsRemoved":0}},"updateStatistics":{"_stats":{"nodesCreated":0,"nodesDeleted":1,"relationshipsCreated":0,"relationshipsDeleted":0,"propertiesSet":0,"labelsAdded":0,"labelsRemoved":0,"indexesAdded":0,"indexesRemoved":0,"constraintsAdded":0,"constraintsRemoved":0}},"plan":false,"profile":false,"notifications":[],"server":{"address":"localhost:7687","version":"Neo4j/3.5.12"},"resultConsumedAfter":{"low":0,"high":0},"resultAvailableAfter":{"low":17,"high":0}}},"error":[],"msg":"Query results"}*/
 const deletePerson = async(req, resp) => {
   let parameters = req.query;
+  if (typeof parameters._id==="undefined" || parameters._id==="") {
+    resp.json({
+      status: false,
+      data: [],
+      error: true,
+      msg: "Please select a valid id to continue.",
+    });
+    return false;
+  }
   let person = new Person({_id: parameters._id});
   let data = await person.delete();
   resp.json({
     status: true,
     data: data,
+    error: [],
+    msg: "Query results",
+  });
+}
+/**
+* @api {delete} /people Delete people
+* @apiName delete people
+* @apiGroup People
+* @apiPermission admin
+*
+* @apiParam {array} _ids The ids of the people for deletion.
+*
+* @apiSuccessExample {json} Success-Response:
+{"status":true,"data":[{"records":[],"summary":{"statement":{"text":"MATCH (n:Event) WHERE id(n)=1149 DELETE n","parameters":{}},"statementType":"w","counters":{"_stats":{"nodesCreated":0,"nodesDeleted":1,"relationshipsCreated":0,"relationshipsDeleted":0,"propertiesSet":0,"labelsAdded":0,"labelsRemoved":0,"indexesAdded":0,"indexesRemoved":0,"constraintsAdded":0,"constraintsRemoved":0}},"updateStatistics":{"_stats":{"nodesCreated":0,"nodesDeleted":1,"relationshipsCreated":0,"relationshipsDeleted":0,"propertiesSet":0,"labelsAdded":0,"labelsRemoved":0,"indexesAdded":0,"indexesRemoved":0,"constraintsAdded":0,"constraintsRemoved":0}},"plan":false,"profile":false,"notifications":[],"server":{"address":"localhost:7687","version":"Neo4j/3.5.12"},"resultConsumedAfter":{"low":0,"high":0},"resultAvailableAfter":{"low":6,"high":0}}}],"error":[],"msg":"Query results"}
+*/
+const deletePeople = async(req, resp) => {
+  let deleteData = req.body;
+  if (typeof deleteData._ids==="undefined" || deleteData._ids.length===0) {
+    resp.json({
+      status: false,
+      data: [],
+      error: true,
+      msg: "Please select valid ids to continue.",
+    });
+    return false;
+  }
+  let responseData = [];
+  for (let i=0; i<deleteData._ids.length; i++) {
+    let _id = deleteData._ids[i];
+    let person = new Person({_id: _id});
+    responseData.push(await person.delete());
+  }
+  resp.json({
+    status: true,
+    data: responseData,
     error: [],
     msg: "Query results",
   });
@@ -477,4 +600,5 @@ module.exports = {
   getPerson: getPerson,
   putPerson: putPerson,
   deletePerson: deletePerson,
+  deletePeople: deletePeople,
 };
